requirements.txt
docx2pdf
# img2pdf
Pillow
PyPDF2
fpdf